+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased - ReleaseDate
 ### Added
-- Reexport `error::Error`.
+- Reexport `error::Error` as `Error`.
 
 ## 0.2.5 - 2021-01-23
 ### Added